+++ module/Photo/src/Photo/Service/Photo.php
@@ -9,6 +9,7 @@ use DateInterval;
 use Decision\Model\Member;
 use Exception;
 use Photo\Model\Hit as HitModel;
+use Photo\Model\Vote as VoteModel;
 use Photo\Model\Tag as TagModel;
 use Photo\Mapper\Tag as TagMapper;
 use Photo\Model\Photo as PhotoModel;
@@ -652,6 +653,24 @@ class Photo extends AbstractAclService
     }
 
     /**
+     * Count a vote for the specified photo.
+     *
+     * @param int $photoId
+     */
+    public function countVote($photoId)
+    {
+        $member = $this->sm->get('user_role')->getMember();
+        if ($this->getVoteMapper()->findVote($photoId, $member->getLidnr()) !== null) {
+            // Already voted
+            return;
+        }
+        $photo = $this->getPhoto($photoId);
+        $vote = new VoteModel($photo, $member);
+        $this->getVoteMapper()->persist($vote);
+        $this->getVoteMapper()->flush();
+    }
+
+    /**
      * Tags a user in the specified photo.
      *
      * @param integer $photoId
@@ -738,6 +757,16 @@ class Photo extends AbstractAclService
     public function getProfilePhotoMapper()
     {
         return $this->sm->get('photo_mapper_profile_photo');
+    }
+
+    /**
+     * Get the vote mapper.
+     *
+     * @return \Photo\Mapper\Vote
+     */
+    public function getVoteMapper()
+    {
+        return $this->sm->get('photo_mapper_vote');
     }
 
     /**