+++ CHANGELOG.md
@@ -1,15 +1,37 @@
+## [v1.6-beta](https://github.com/GEWIS/gewisweb/tree/v1.6-beta) (2019-11-02)
+
+* Implemented a new photo viewer next to the old photo viewer
+* Improved the jobs overview
+* Improved user registration
+* Improved exception logging
+* Various UI improvements
+* Various bugfixes
+
+---
+
+## [v1.5.3](https://github.com/GEWIS/gewisweb/tree/v1.5.3) (2019-09-26)
+
+* Various UI improvements
+* Various bugfixes
+
+---
+
 ## [v1.5.2](https://github.com/GEWIS/gewisweb/tree/v1.5.2) (2019-09-19)
 
 * Improved option calendar
 * Used more informative title for emails to GEFLITST
 * Various bugfixes
 
-## [v1.5.1](https://github.com/GEWIS/gewisweb/tree/v1.5.1) (2019-08-31)
+---
+
+## [v1.5.1](https://github.com/GEWIS/gewisweb/tree/v1.5.1) (2019-09-02)
 
 * Implemented RFC2324
 * Updated dependencies
 * Various UI improvements
 * Various bugfixes
+
+---
 
 ## [v1.5.0](https://github.com/GEWIS/gewisweb/tree/v1.5.0) (2019-06-25)
 
@@ -18,6 +40,8 @@
 * Added social media references
 * Various UI improvements
 * Various bugfixes
+
+---
 
 ## [v1.4.2](https://github.com/GEWIS/gewisweb/tree/v1.4.2) (2019-05-10)
 
@@ -28,6 +52,8 @@
 * Various UI improvements
 * Various bugfixes
 
+---
+
 ## [v1.4.1](https://github.com/GEWIS/gewisweb/tree/v1.4.1) (2019-03-21)
 
 * Sort meeting documents on number
@@ -35,6 +61,8 @@
 * Add overview of users subscribed activities
 * Various UI improvements
 * Various bugfixes
+
+---
 
 ## [v1.4.0](https://github.com/GEWIS/gewisweb/tree/v1.4.0) (2019-02-06)
 
@@ -44,6 +72,8 @@
 * Disclaimer on educational material
 * Various UI improvements
 * Various bugfixes
+
+---
 
 ## [v1.3.2](https://github.com/GEWIS/gewisweb/tree/v1.3.2) (2018-03-02)
 
@@ -123,5 +153,3 @@
 ## [v1.0](https://github.com/GEWIS/gewisweb/tree/v1.0) (2016-02-06)
 
 * Initial release, lots of great features
-
----